[CI] Avoid using EXTRA_INDEX_URL in pip install (#1424)
While not currently a vulnerability, using this option can introduce
risks of supply chain attacks, where an attacker adds a malicious
package into the default public repository with the same name as one
that you are installing. The best practice is to instead use --index-url
when a package is not available in the default repo, so I've updated to
use this option instead.

For full disclosure, this is currently causing Microsoft's internal
security checks to fail when building triton (which is why I care about
this theoretical issue/best practice).