Fixes a bug in `ConvertLayoutOp` canonicalization patterns. (#2671)
In `TritonGPU` phase, under `CanonicalizeConvertFromConvert` pattern,
`cvt(view) -> cvt` always runs as long as the `view` is consider as
inexpensive, regardless of the encodings of the operand/result.

In `TritonGPUToLLVM` phase, `ViewOp` is converted to unpacking and
packing operations, which requires the element type to match between
unpacking and packing. However, different encodings might use different
element types, which makes it unsafe for `view` to change encoding
unconditionally.