[TESTS] Allow user defined output_dtype and acc_dtype in matmul tests (#2769)
In this PR we are allowing to manually set acc_dtype and output_dtype in
matmul test.

They are:
- `acc_dtype`: So users of the test class can specify the type used
internally in the dot, and not the one set by default given the two
types. There are several restrictions for these types anyway.
- `output_dtype`: The return type of the matmul. I included a few tests
in the case of making a dot with two float16.

I also had to modify test_matmul to use a small range of values to
prevent numerical issues. In the case of testing with two `float16` and
`acc_dtype` `float16`, since I can't force torch to use `float16`
internally (it uses `float32`), I was having precision issues when
comparing the results with triton.