[FRONTEND][BACKEND] Add tl.histogram op (#2890)
Add a new op to compute the histogram of a tensor. The op takes a 1D
tensor and return a new tensor also 1D where each element represents 1
bin. Bins start at 0 and are of width 1. A more generic histogram can
easily be implemented on top of this by normalizing the inputs.

A lot of the credits goes to @apgoucher for writing the algorithm to do
histogram efficiently within a warp.