[FRONTEND] Fix triton.language.dtype repr (#3342)
While working on adding `triton.language.dtype` to PyTorch, I discovered
that `triton.language.float32` emits `triton.language.fp32` which I
assume is done in order to be consistent with numpy and torch; however,
this results in an un-evaluable expression as `triton.language.fp32`
does not correspond to anything.
I thought of adding aliases to float32 as fp32 but that meant
duplicating bunch of dtypes. This solution seems cleaner.