[backend][amd] Switch back to code object v4 and be consistent (#3515)
https://github.com/openai/triton/pull/3493 moved the AMD backend to use
code object v5. But that reveals some issues that we still need to
investigate and resolve. For now switch back to v4.

Though we need to be consistent by attaching the LLVM module attribute
`amdhsa_code_object_version`, given that's what the AMDGPU LLVM backend
is trying to query:

https://github.com/llvm/llvm-project/blob/6f44bb77/llvm/lib/Target/AMDGPU/Utils/AMDGPUBaseInfo.cpp#L166-L173
If missing it will fallback to use v5, which was happening before
https://github.com/openai/triton/pull/3493.

And FWIW, the magic string used to be `amdgpu_code_object_version`:
https://github.com/llvm/llvm-project/pull/79905. The last LLVM bump
https://github.com/openai/triton/pull/3463 picked up the renaming patch.