[OPTIMIZER] Take numWarps into account for  Hopper mma op (#2956)
This changes the `wgmma` instruction shape based on the total number of
warps.

Instead of always using the largest version of `wgmma`, it honors
the user's `numWarps` hint, and uses a smaller `wgmma` shape to
distribute the work to all warps, rather than having some of them
idle.

Using m's shape, it calculates how many warps will be used in the m
dimension, then see how many are left for the n dimension. Then, it
chooses the largest N such that it is still evenly distributed.
    
This resolves issue https://github.com/openai/triton/issues/2662.