[Backend] Only consider valid LoadOps for pipelining (#3517)
This PR fixes #3454

Previously, we included all LoadOp(s) into opInfo, whose contents are
used for pipelining. However, this approach causes issues when we decide
to not pipeline some LoadOps. Particularly, those LoadOps that are
excluded will end up with an inconsistent state which violates the
relevant integrity checks, and cause ICEs.

We fix the issue by only including LoadOps that are valid for
pipelining.