[backend][amd] Add support for device assert (#3552)
This commit adds support for device assert in the AMD
backend. This creates a new virtual method on `TargetInfo`
so that we can specialize for different backends, given
that on AMD side we don't have the same lowering flow:
right now `__assert_fail` is not supported in the AMDGPU
backend; it's in the runtime. 


https://github.com/ROCm/clr/commit/7137a296ddab93708a245955cd6a7357922975cc

Instead we need to layer it on top of print and do trap
afterwards.