[frontend] Improve function def parsing with stacked decorators. (#3564)
Under certain conditions where we apply another decorator on top of
triton.jit, there is a small chance that in the decorator body there's a
"def" substring in it. This is problematic when JITFunction parses
function def using str.find('def'). This causes an issue compiling some
pytorch model when we happened to decorate a jit function with a hash
string argument which has "def" in it.

This diff tries a bit harder to locate the real "def" position using a
more complex regex which should further reduce the chance to hit this
error.