[Tests] Add test for local_load to #mma layout. (#3569)
This is a TTGIR test for the bug worked around in PR
https://github.com/openai/triton/pull/3561.

While we're here, we also do a minor cleanup in test_core.py.

A previous version of this patch needed us to fix the regex for parsing
an MLIR
file to find the individual arguments.  Previously it would include the
function's closing paren in the argument type. We keep this because it's
a
reasonable fix.

I am extremely frightened by this test; if you comment out the skip() at
the
beginning, many tests fail due to OOB memory accesses or wrong results,
in
addition to the asserts added in PR
https://github.com/openai/triton/pull/3561. Who knows if we hit this
today?