[AMD backend] fix test reproducer (#3642)
This PR is to fix the unit test `test_reproducer.py`. 

Reason of the failure is that compiler in amd backend calls the function
`pm.run(mod)` multiple times, reproducer is only for the passes in the
last call of the function `pm.run(mod)`, so ttgir->llvm pass is not
included in the last call.

Fix is to call the `pm.run(mod)` only once in the function
`make_llir()`.