[FRONTEND] add experimental descriptor load mapping to TMA (#3739)
This adds an experimental escape hatch to access to TMA operations
explicitly through this front end. This is meant for experiments only as
it breaks portability. This will allow exercising TMA operations short
term and will be removed after that.

Users should not rely on this feature in their production code.