[backend][nvidia] move NVGPU dialect to third_party/nvidia/ (#3773)
This commit moves the `NVGPU` dialect to the `third_party/nvidia`
directory. To make it happen we need to clear some unnecessary
dependency on `NVGPU` dialect in the AMD backend and move
some utilities into the `third_party/nvidia` directory too. Overall this
gives us a better structure and creates a stricter boundary between
core and backends.