[BACKEND] Bringing back reverted change and limiting the hack to ReduceOp (#3768)
Original change:
[BACKEND] Adjusting the hasConvertToMMATransisitiveUse to look through if results (#3694)

hasConvertToMMATransisitiveUse is supposed to decide if it is a
beneficial to propagate given anchor layout to other ops. The
propagation code was missing handling of cases where the def-use chain
was going through the `if` results inside the loop, which was resulting
in sub-optimal code in some cases.