Handle the case where env var is string (#3833)
Summary: Previously we assume all env vars are bool. This diff changes
the metadata envVars from map<string, bool> to map<string, string>. If
the value is a bool value, the entry in envVars will be 'true' or
'false', otherwise, it will be the string value. If the value is
empty/not specified, it will not show.

---------

Co-authored-by: Manman Ren <mren@fb.com>