Add parameter `interleave` to the `warpReduce` interface of `TargetInfo` which was missed. (#4041)
The reduce lowering pattern, ReduceOpConversion, calls the backend
target-specific interface to generate a high-performance operation for
horizontal warp reduction across lanes. If the backend target doesn't
support this, it falls back to the common horizontal warp reduction with
the ShuffleXor operation, which might be suboptimal.

There is a parameter interleave in warpReduce that is not defined in the
TargetInfoBase interface warpReduce.

The backend may generate incorrect code without information about how to
reduce the value across lanes with an interleave stride.

This issue doesn't impact the in-tree backends, AMD and NV GPU.

- The AMD backend always uses ShuffleXor for horizontal warp reduction.
- The NV backend only generates high-performance horizontal warp
reduction when numLaneToReduce equals 32. (The interleave is always 1
when numLaneToReduce is 32 on NV, where the warp size is always 32.)
 
This issue impacts the out-of-tree backend, Intel GPU.

- The Intel GPU supports various high-performance horizontal warp
reduction operations for different numLaneToReduce and interleave
values.

To fix this, we add the interleave parameter to the TargetInfo's
warpReduce method and use the same interleave value as when generating
the horizontal reduction with ShuffleXor.