[OPTIMIZER] Fix layout error after backward remat (#4311)
Fixes pytorch/pytorch#130101

In the linked issue `getConvertBackwardSlice` is returning a slice that
requires two different layouts for the same operation, which isn't
supported. There is a guard that checks `layouts.find(currentValue)` and
bails out when this doesn't match, however this check came after the
`visited` check so we would skip the queue item before we ever got to
the check.

To fix this I change the visited check to only apply if the current
`(value, encoding)` pair has been enqueued previously and apply this
uniformly to everywhere we want to update a layout.

The reproducer is admittedly long, but I was able to reduce it down to
the exact slice that was triggering the failure.