Build LLVMAarch64CodeGen if CMAKE_OSX_ARCHITECTURES is arm64. (#4637)
I'm trying to make arm64 Triton builds work. I've found that on my
machine, everything works if I make this change and build with
CMAKE_OSX_ARCHITECTURES=arm64.
    
I'm not totally sure why this change is needed. From talking to chatgpt,
it should be sufficient to run the build with `arch -arm64`; that should
set CMAKE_SYSTEM_PROCESSOR. But...it doesn't. And you are not allowed to
override this variable from the command-line.
    
Anyway this seems to work, and who among us can say they truly
understand cmake?