[BACKEND] Linear Layout with `stmatrix` part 1: support `stmatrix` for `convert_layout` ops (#4627)
The PR has initiated the use of linear layout in `convert_layout`
operations when the source operand can be stored using `stmatrix`.

Please note that we are still using the legacy code path for
`local_alloc` operations when `stmatrix` could be used. This will be
addressed in the next PR (part 2).