[ANALYSIS] Don't consider descending sequences as contiguous in AxisInfoAnalysis (#4871)
Contiguity is used to issue wide load operations instead of multiple
loads.
This always assumes that the address of the first element in a sequence
can
be used to load the sequence of elements. If a sequence is diminishing,
it
leads to a wrong wide load operation.

This patch fixes that by not preserving a contiguity of RHS for SubIOp
operation.

- [x] I am not making a trivial change, such as fixing a typo in a
comment.

- [x] I have written a PR description following these
  [rules](https://cbea.ms/git-commit/#why-not-how).

- [ ] I have run `pre-commit run --from-ref origin/main --to-ref HEAD`.

- Select one of the following.
  - [x] I have added tests.
    - `/test` for `lit` tests
    - `/unittest` for C++ tests
    - `/python/test` for end-to-end tests
  - [ ] This PR does not need a test because `FILL THIS IN`.

- Select one of the following.
  - [x] I have not added any `lit` tests.
- [ ] The `lit` tests I have added follow these [best
practices](https://mlir.llvm.org/getting_started/TestingGuide/#filecheck-best-practices),
including the "tests should be minimal" section. (Usually running Python
code
    and using the instructions it generates is not minimal.)

---------

Signed-off-by: Ilya Enkovich <ilya.enkovich@intel.com>